Refactored and Integrated both Trees
In this commit, I have integrated the struct "Point" to work with K-Dimensional Trees, hence making the operative system for both the trees to be common.

I also simplified and refactored majority of the code so as to make it more understandable and comprehensive.

I introduced some new functions like subtractPoints, dotProduct, etc. to facilitate the calculation of points as vectors, which are an essential requirement of the buildBallTree function.

I have also changed the functionality of the buildBallTree function as it was inconsistent in choosing the points becuase it separating the points based on radius - inefficient and incorrect way of implementation.

At last, I have re-set the main function and will generate the trees in the next commit. I also have removed comments in this iteration because refactoring the code with comments was a bit overwhelming. I will add comments in the next commit!